ms-customer-info/src/main/java/com/customer/controller/CustomerController.java
package com.customer.controller;

import com.customer.controller.dto.CustomerResponseDto;
import com.customer.service.CustomerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import java.util.List;

@RestController
@RequestMapping("/api/customers")
public class CustomerController {

    @Autowired
    private CustomerService customerService;

    @GetMapping("/{id}")
    public ResponseEntity<CustomerResponseDto> getCustomerById(@PathVariable Long id) {
        CustomerResponseDto responseDto = customerService.getCustomerById(id);
        return ResponseEntity.ok(responseDto);
    }

    @GetMapping
    public ResponseEntity<List<CustomerResponseDto>> getAllCustomers() {
        List<CustomerResponseDto> responseDtoList = customerService.getAllCustomers();
        return ResponseEntity.ok(responseDtoList);
    }
}